add more test code
