+++ segmenter_test.go
@@ -288,46 +288,6 @@ func TestLoadDictionary(t *testing.T) {
 	tt.False(t, ok)
 }
 
-func TestHMM(t *testing.T) {
-	prodSeg.LoadDict()
-
-	hmm := prodSeg.HMMCutMod("纽约时代广场")
-	tt.Equal(t, 2, len(hmm))
-	tt.Equal(t, "纽约", hmm[0])
-	tt.Equal(t, "时代广场", hmm[1])
-
-	text := "纽约时代广场, 纽约帝国大厦, 旧金山湾金门大桥"
-	tx := prodSeg.Cut(text, true)
-	tt.Equal(t, 7, len(tx))
-	tt.Equal(t, "[纽约时代广场 ,  纽约 帝国大厦 ,  旧金山湾 金门大桥]", tx)
-
-	tx = prodSeg.cutDAGNoHMM(text)
-	tt.Equal(t, 9, len(tx))
-	tt.Equal(t, "[纽约时代广场 ,   纽约 帝国大厦 ,   旧金山湾 金门大桥]", tx)
-
-	tx = prodSeg.Trim(tx)
-	tt.Equal(t, 5, len(tx))
-	tt.Equal(t, "[纽约时代广场 纽约 帝国大厦 旧金山湾 金门大桥]", tx)
-
-	tx = prodSeg.CutAll(text)
-	tt.Equal(t, 21, len(tx))
-	tt.Equal(t,
-		"[纽约 纽约时代广场 时代 时代广场 广场 ,   纽约 帝国 帝国大厦 国大 大厦 ,   旧金山 旧金山湾 金山 山湾 金门 金门大桥 大桥]",
-		tx)
-
-	tx = prodSeg.CutSearch(text, false)
-	tt.Equal(t, 20, len(tx))
-	tt.Equal(t,
-		"[纽约 时代 广场 纽约时代广场 ,   纽约 帝国 国大 大厦 帝国大厦 ,   金山 山湾 旧金山 旧金山湾 金门 大桥 金门大桥]",
-		tx)
-
-	tx = prodSeg.CutSearch(text, true)
-	tt.Equal(t, 18, len(tx))
-	tt.Equal(t,
-		"[纽约 时代 广场 纽约时代广场 ,  纽约 帝国 国大 大厦 帝国大厦 ,  金山 山湾 旧金山 旧金山湾 金门 大桥 金门大桥]",
-		tx)
-}
-
 func TestInAlphaNum(t *testing.T) {
 	// var seg Segmenter
 	// AlphaNum = true
